search of location added

+++ WeatherApp/JSON Classes/Main.cs
@@ -8,10 +8,10 @@ namespace WeatherApp.JSON_Classes
 {
     public class Main
     {
-        public int temp { get; set; }
+        public double temp { get; set; }
         public int pressure { get; set; }
         public int humidity { get; set; }
-        public int temp_min { get; set; }
-        public int temp_max { get; set; }
+        public double temp_min { get; set; }
+        public double temp_max { get; set; }
     }
 }
